+++ src/pages/settings/features/support/SupportFeatures.tsx
@@ -1,9 +1,9 @@
 import React, { useState } from "react"
 import { Helmet } from "react-helmet"
-import BreadCrumbs from "../../../components/settings/BreadCrumbs"
-import Header from "../../../components/settings/Header"
-import ConstantsRegistry from "../../../global/ConstantsRegistry"
-import { supportFeaturesRoutes } from "../../../routes/settings/featuresRoutes"
+import BreadCrumbs from "../../../../components/settings/BreadCrumbs"
+import Header from "../../../../components/settings/Header"
+import ConstantsRegistry from "../../../../global/ConstantsRegistry"
+import { supportFeaturesRoutes } from "../../../../routes/settings/featuresRoutes"
 import AgentFeatures from "./AgentFeatures"
 import AnnouncementFeatures from "./AnnouncementFeatures"
 import ClientFeatures from "./ClientFeatures"
