+++ src/routes/settings/featuresRoutes.tsx
@@ -1,4 +1,4 @@
-import SupportFeatures from "../../pages/settings/features/SupportFeatures";
+import SupportFeatures from "../../pages/settings/features/support/SupportFeatures";
 
 export const supportFeaturesRoutes = [
     { path: "/a/default/settings/features/support", element: <SupportFeatures />, activeMenu: 'support' },
